+++ R/simpaper.R
@@ -3,7 +3,7 @@
 #' An uncoupled two-dimensional Ornstein-Uhlenbeck process simulated with `affectOU`.
 #'
 #' The process is defined by the stochastic differential equation:
-#' \deqn{d\mathbf{X}_t = \mathbf{\Theta}(\mathbf{\mu} - \mathbf{X}_t)dt + \mathbf{\Gamma}^{1/2}d\mathbf{W}_t}
+#' \deqn{d\mathbf{X}_t = \mathbf{\Theta}(\mathbf{\mu} - \mathbf{X}_t)dt + \mathbf{\Gamma}d\mathbf{W}_t}
 #'
 #' where:
 #' \itemize{